mqttimport: replace timeout with self syncing
This commit replaces the dumb timeout method
with a more elegant self-syncing method to detect
abscense of a specific topic.
This self-syncing sends at last in the init phase a specific topic.
It is assumed here that this will be received after all other
subscribed topics have been received, so no more items will follow.

Topics are kept in the local linked list until we did actually
receive it, either because it existed already, or we send it ourselve.
Since the empty list is the criterium to exit,
this fixes queueing when importing a large set of items.
Before this commit, it was all queued, and the items were destroyed,
so the program already thought it could exit.